# docs/openapi.yaml
openapi: 3.1.0
info:
  title: zkAPI v2
  version: 2.0.0-testnet
  description: Local compatibility API plus the private Groth16 protocol API.
paths:
...
  /v2/requests:
    post:
      summary: Submit an exact-payload-bound Groth16 request proof
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/ApiRequestV2"}
      responses:
        "200": {description: Provider result and signed next state}
        "400": {description: Invalid request or proof}
        "409": {description: Replayed nullifier}
  /v2/openrouter/leases:
    post:
      summary: Open a prompt-private, spending-limited OpenRouter lease
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/ApiRequestV2"}
      responses:
        "201":
          description: One-time plaintext runtime key and lease timing
          content:
            application/json:
              schema: {$ref: "#/components/schemas/OpenRouterLease"}
        "400": {description: Invalid request, proof, or disabled lease mode}
        "409": {description: Existing active or settling lease}
  /v2/openrouter/leases/{client_request_id}:
    get:
      summary: Read non-secret lease state; never returns the runtime key
      parameters:
        - in: path
          name: client_request_id
          required: true
          schema: {type: string}
      responses:
        "200": {description: Lease timing, status, and finalized aggregate usage}
        "404": {description: Unknown lease}
...
  /v2/requests/{client_request_id}:
    get:
      summary: Recover a request transcript
      parameters:
        - in: path
          name: client_request_id
          required: true
          schema: {type: string}
      responses:
        "200": {description: Finalized transcript}
        "404": {description: Unknown request}
components:
  schemas:
    Groth16Proof:
      type: object
      required: [backend, proof]
      properties:
        backend: {const: groth16_bn254}
        proof: {type: string, contentEncoding: base64}
    ApiRequestV2:
      type: object
      required: [client_request_id, payload, payload_hash, public_inputs, proof]
      properties:
        client_request_id: {type: string}
        payload: {type: string}
        payload_hash: {type: string, pattern: "^0x[0-9a-f]+$"}
        public_inputs: {type: object, additionalProperties: true}
        proof: {$ref: "#/components/schemas/Groth16Proof"}
    OpenRouterLease:
      type: object
      required: [status, client_request_id, api_key, openrouter_api_base, issued_at, expires_at, valid_for_seconds, settle_after, spending_limit_usd]
      properties:
        status: {const: active}
        client_request_id: {type: string}
        api_key: {type: string, description: Returned once and never persisted by zkapi-serverd}
        openrouter_api_base: {type: string, format: uri}
        issued_at: {type: integer}
        expires_at: {type: integer}
        valid_for_seconds: {type: integer}
        settle_after: {type: integer}
        spending_limit_usd: {type: number}